Update content.inc.php
If the 'tv' variable's value has one or more equal sign, the string is cut at the first equal sign. After this change, you can use equal sign in the value of 'tv' variable. It is very useful, when you add a special character or icon ( for example: <span class="glyphicons glyphicons-envelope"></span>) for the value. It works after this modification.